requirements.txt
streamlit
opencv-python
tensorflow
keras
Pillow